tools/requirements.txt
openai>=1.0
docling
python-dotenv
